Enhance the bbq page and close a bunch of issues.
This change allows for the complete project lifecycle to take place on
the BBQ page. The coffeescript was refactored to use
javascript objects, and we can create a proposal, promote it to a
project, go live, edit, and shut down.

This change also fixes the test for when projects and proposals no
longer have slug fields, so there are two lines commented
 out in models/project.py that refer to OF-64.

Tickets affected by this change:
 - OF-25: Projects service (closed)
 - OF-60: BBQ Page object lifecycle (closed)
 - OF-62: Proposal Service (closed)
 - OF-64: Remove slug, use ID based keys (pending a branch rebase)
 - OF-68: Remove add_media from project service (closed)

Note that all but one of the tests are expected to pass.