Midway checkin for new Media service (OF-75, OF-76, OF-77)
This change adds the basics for uploading to the blobstore and saving
the assets. There are also various small fixes and some commented ou
t code in order to make things work.

Currently, the Content Editor works to upload an image if you change the
endpoint call in AppToolsJS to be:
 - $.apptools.api.media.generate_endpoint().fulfill
 - and also change the response that is used from "response.url" to
   "response.endpoints[0]"

I attempted to write some basic tests for the blobstore but have not
been successful in uploading a file during testing.

Also in this change:
 - Add media service to config file.
 - Add pending flag to assets and updated the fixtures.
 - Added an Image record type to store images on projects.
 - Skeleton for some other media api calls, see the wiki page for more.
 - Added tabs and uploader widget to the cakefile.
 - Fixed the app.admin/base.admin js mix up.

Note that I commented out some debug statements that were being shown
during tests.