Working image uploads, as Data URL types. (OF-66)
This checkin allows admins to drop images onto projects on the bbq page
and have them attach and be served for that project.

Eventually we would like to change it so that we upload images as raw
data instead of data urls, but this works for now.

Also:
 - Moved the decrypt and encrypt functions higher up so they are on
   handlers and services.
 - Encrypt a few project keys.

Note: This checkin requires that AppToolsJS use readAsDataURL instead of
readAsBinaryString.