Proposal page wizard update [OF-214]
This change updates the proposal wizard to the new data model and makes
sure it works.

Now you can create a proposal on the propose page and then accept it
with a call to api.proposal.promote with the key. Then you'll
have a new project page to work with at the custom url you put in the
proposal.