Social Icons & Project Page Fixes
This commit fixes the layout issues experienced on our project page, and
ensures
the sidebar is floating & sticking as should be. This commit also adds
four new
font files, called 'momentron-social', that contain our social icon
typefaces.
This is a partial commit to unblock Ethan from working on the project
page.

Other changes:
 - attendant bugfixes for apptoolsJS to ensure consistent widget action
   with new
   util.coffee setup
 - Minor modifications to _fonts.sass & _forms.sass