Centralize proposal home template code
Seems like we ended up duplicating the proposal home template code
somehow. This change clears that up by condensing it all into the
layout/propsal.html version, since that is the one everyone seems to be
using the most. This clears the way for me to update the
proposal sidebar.

Also in this change:
 - Show the proper status of a project on it's page and only allow
   donating in the 'open' phase.
 - Added backend service code for submitting a proposal. Not sure if
   this was somehow lost or just never written.
 - Allow the proposal title to be any text for right now.
 - Fix some js errors when buttons don't exist.