Show more realistic data in the homepage graphs
These are the new graphs on the homepage (powered by NVD3, which is
added in this change):
 1. Time series data for number of donations and total donation amount.
 2. Pie chart showing where donations are going in the last week.
 3. Time series showing number of projects and number of funded
projects.

And then there is one more box with just numbers, including:
 1. Total number of projects.
 2. Total donations site-wide.
 3. Number of individual donations site-wide.

Note that all the data being displayed is still fake, but it should be
easy to drop in real data when we are ready (low priority).